% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{arithmetic}
\alias{arithmetic}
\title{Teaching Arithmetic Data}
\format{
An object of class \code{data.frame} with 45 rows and 2 columns.
}
\source{
Unpublished lecture notes, Imperial College, London. Extracted from the SMPractical package and distributed under GPL-3 license
}
\usage{
arithmetic
}
\description{
45 school pupils were divided at random into five groups of size 9.
Groups A (\code{control 1}) and B (\code{control 2}) were taught arithmetic in
separate classes by the usual method. Groups C (\code{praise}), D (\code{reprove}),
and E (\code{ignore}) were taught together for several days.  On each day group
C were publically praised, group D were publically reproved, and group E
were ignored.  The responses are from a standard test taken by all pupils at
the end of the period.

A data frame with 45 rows and 2 variables:
\describe{
\item{group}{[factor] experimental group, one of \code{control 1}, \code{control 2}, \code{praise}, \code{reprove} or \code{ignore}}
\item{score}{[numeric] score on arithmetic test}
}
}
\examples{

data(arithmetic)
anova(aov(score ~ group, data = arithmetic))
}
\references{
Davison, A. C. (2003) \emph{Statistical Models}.  Cambridge
University Press. Page 427.
}
\keyword{datasets}
