% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{AVC02}
\alias{AVC02}
\title{Anandarajan, Viger and Curatola (2002)}
\format{
A data frame with 132 rows and 4 variables:
\describe{
  \item{\code{format}}{[factor] disclosure format, one of \code{integrated note}, \code{stand-alone note} or \code{modified auditor report}}
  \item{\code{prime}}{[double] rating for the interest rate premium assessed}
  \item{\code{debt}}{[integer] rating for the ability to service debt}
  \item{\code{profitability}}{[integer] rating for the ability to improve profitability}
}
<<<<<<< HEAD
}
\source{
Leo Belzile
}
\usage{
AVC02
}
\description{
A study on the impact of communication means on perceived risk of organization on the brink of bankrupcy vehiculated through different disclosure format.
}
\details{
This data was simulated based on the summary statistics presented in the paper
}
\references{
Anandarajan, A., Viger, C., & Curatola, A. P. (2002). \emph{An experimental investigation of alternative going-concern reporting formats: A Canadian experience.} Canadian Accounting Perspectives, 1(\bold{2}), 141-162. \url{https://doi.org/10.1506/5947-NQTC-C3Y5-H46N}
=======

A data frame with 132 rows and 4 variables:
\describe{
  \item{\code{format}}{[factor] disclosure format, one of \code{integrated note}, \code{stand-alone note} or \code{modified auditor report}}
  \item{\code{prime}}{[double] rating for the interest rate premium assessed}
  \item{\code{debt}}{[integer] rating for the ability to service debt}
  \item{\code{profitability}}{[integer] rating for the ability to improve profitability}
}
}
\source{
Leo Belzile

Leo Belzile
}
\usage{
AVC02

AVC02
}
\description{
A study on the impact of communication means on perceived risk of organization on the brink of bankrupcy vehiculated through different disclosure format.

A study on the impact of communication means on perceived risk of organization on the brink of bankrupcy vehiculated through different disclosure format.
}
\details{
This data was simulated based on the summary statistics presented in the paper

This data was simulated based on the summary statistics presented in the paper
}
\references{
Anandarajan, A., Viger, C., & Curatola, A. P. (2002). \emph{An experimental investigation of alternative going-concern reporting formats: A Canadian experience.} Canadian Accounting Perspectives, 1(\bold{2}), 141-162. \url{https://doi.org/10.1506/5947-NQTC-C3Y5-H46N}

Anandarajan, A., Viger, C., & Curatola, A. P. (2002). \emph{An experimental investigation of alternative going-concern reporting formats: A Canadian experience.} Canadian Accounting Perspectives, 1(\bold{2}), 141-162. \url{https://doi.org/10.1506/5947-NQTC-C3Y5-H46N}
>>>>>>> update2
}
\keyword{datasets}
