% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BJF14_S1}
\alias{BJF14_S1}
\title{Bastian, Jetten and Ferris (2014)}
\format{
A data frame with 54 rows and 10 variables:
\describe{
\item{\code{bonding}}{[double] score for the perceived bonding with strangers, average of Likert scales}
\item{\code{condition}}{[factor] experimental condition, either \code{Control} or \code{Pain}}
\item{\code{pain}}{Likert scale rating for the amount of pain, 'How intense was the pain you experienced?', ranging from not at all painful (1) to intensely painful (10)}
\item{\code{unpleasant}}{Likert scale rating for unpleasanteness, 'How unpleasant was the pain you experienced?' ranging from not at all (1) to the most intense bad feeling imaginable (10)}
\item{\code{positive}}{PANAS score for positive affect}
\item{\code{negative}}{PANAS score for negative affect}
\item{\code{threat}}{average ALES subscale for the perception of the physical task}
\item{\code{challenge}}{average ALES subscale for the perception of the physical task}
\item{\code{groupsize}}{size of group during experiment}
\item{\code{age}}{age of participant}
\item{\code{gender}}{[factor] gender, either male or female}
}
}
\source{
Open Science Foundation, via \url{https://osf.io/9k3sw/}, unspecified license
}
\usage{
BJF14_S1
}
\description{
Data for Study 1 of Bastian, Jetten and Ferris (2014) social effects of pain. Study 1 contains a one-way analysis of variance, a series of t-test, a mediation analysis and some linear regression.

The response, \code{bonding}, is the average of 7 questions measured on a seven point scale, ranging from strongly disagree (1) to strongly agree (7).
}
\references{
Bastian, B., Jetten, J., & Ferris, L. J. (2014). \emph{Pain as Social Glue: Shared Pain Increases Cooperation}. Psychological Science, 25(\bold{11}), 2079–2085. \doi{10.1177/0956797614545886}
}
\keyword{datasets}
