% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BL22_E}
\alias{BL22_E}
\title{Brucks and Levav (2022), eyetracker data}
\format{

}
\source{
Research Box 282, \url{https://researchbox.org/282}, unspecified license
}
\usage{
BL22_E
}
\description{
Brucks and Levav (2022), eyetracker data
}
\details{
The authors measured via an eyetracker the time spent gazing at the room, at the other partner and at the task during their lab experiment
}
\examples{
\dontrun{
library(lme4)
mod <- lmer(task_time ~ cond + (1|id),
            data= BL22_E)
summary(mod)
}
}
\references{
Brucks, M.S. and J. Levav (2022)\emph{Virtual communication curbs creative idea generation}. Nature 605, 108–112. \url{https://doi.org/10.1038/s41586-022-04643-y}
}
\keyword{datasets}
