% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BL22_L}
\alias{BL22_L}
\title{Brucks and Levav (2022), Lab experiment}
\format{
A data frame with 301 rows and 8 variables:
\describe{
  \item{\code{cond}}{[factor] experimental condition, either \code{video} or \code{f2f} for face-to-face}
  \item{\code{nidea}}{[integer] number of ideas generated by pair}
  \item{\code{ncreative}}{[integer] number of ideas that were above the average creativity score}
  \item{\code{select_creative}}{[double] creativity score of the idea each group selected as their most creative idea}
  \item{\code{top_creative}}{[double] creativity score of the highest scoring idea in each group}
  \item{\code{avg_creative}}{[double] average creativity of all ideas generated by each group}
  \item{\code{study}}{[factor] study number}
}
}
\source{
Research Box 282, \url{https://researchbox.org/282}, unspecified license
}
\usage{
BL22_L
}
\description{
Brucks and Levav (2022), Lab experiment
}
\details{
The authors recruted 301 pairs of people and assigned them to either face-to-face interactions or videocall. They measured the number of ideas generated during the discussion as well as the fraction of those which were deemed creative using a `creativity_score`. The latter was measured by two undergraduates who evaluated all ideas based on novelty using a Likert scale with 7 points. Participants were asked to pick their most novel idea: the `diff` column measures the  and a `decision_error` score (with smaller values of the latter meaning better decisions).
}
\references{
Brucks, M.S. and J. Levav (2022)\emph{Virtual communication curbs creative idea generation}. Nature 605, 108–112. \url{https://doi.org/10.1038/s41586-022-04643-y}
}
\keyword{datasets}
