% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BM04_T2}
\alias{BM04_T2}
\title{Bertrand and Mullainathan (2004), Table 2}
\format{
a data frame in long format of 9 observations on the following variables
\describe{
\item{\code{black}}{number of black partipants called back, out of two}
\item{\code{white}}{number of white partipants called back, out of two}
\item{\code{count}}{[integer] count of occurence of each combination}
}
}
\usage{
BM04_T2
}
\description{
Bertrand and Mullainathan (2004), Table 2
}
\references{
Bertrand, M., and S. Mullainathan. 2004. \emph{Are Emily and Greg More Employable Than Lakisha and Jamal? A Field Experiment on Labor Market Discrimination.} American Economic Review 94 (\bold{4}): 991-1013. \doi{10.1257/0002828042002561}.
}
\keyword{datasets}
