% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BMH19_S2}
\alias{BMH19_S2}
\title{Bobak, Mileva and Hancock (2019), Study 2}
\format{
A data frame with 156 rows and 5 variables:
\describe{
  \item{\code{id}}{[integer] participant ID}
  \item{\code{gender}}{[factor] gender of participant, either \code{male} or \code{female}}
  \item{\code{age}}{[integer] age of participant}
  \item{\code{color}}{[factor] color of photographs, either grayscale (\code{mon}), colored (\code{col}) or mixed (\code{mix})}
  \item{\code{pcorr}}{[double] percentage of correctly detected match}
}
}
\source{
Supplementary material, article licensed under CC BY 4.0
}
\usage{
BMH19_S2
}
\description{
Data on matching of pairs of pictures from official ID depending on the color.
Each of the 52 participants was assigned pictures of celebrities and needed to decide whether pairs of picture (potentially mixed or both monochrome or color) were of the same person.
}
\references{
Bobak, A. K., Mileva, V. R., & Hancock, P. J. B. (2019). \emph{A grey area: How does image hue affect unfamiliar face matching?} Cognitive Research: Principles and Implications, 4(\bold{1}), 27. \doi{10.1186/s41235-019-0174-3}
}
\keyword{datasets}
