% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BRLS21_EDA}
\alias{BRLS21_EDA}
\title{Brodeur et al. (2021), Table 3}
\format{
A data frame with 10942 rows and 6 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{reltime}}{[integer] relative time of the task}
  \item{\code{task}}{[factor] task, one of \code{baseline} or distraction, either a notification from the \code{cell}, from a phone call \code{voicecall} or via the smartwatch (\code{watch})}
  \item{\code{occurence}}{[integer] number of occurences}
  \item{\code{eda}}{[double] electrodermal activity measure}
  \item{\code{hr}}{[double] heart rate}
}
}
\source{
Shang Lin Chen, personal communication, distributed under CC BY-NC-SA 4.0
}
\usage{
BRLS21_EDA
}
\description{
Electrodermal activity measured by a smartwatch as part of an experiment on
impact of smartwatches on distraction while driving. The experiment was conducted at Tech3Lab, but the data are not utilized in the paper.
}
\references{
Brodeur, M., Ruer, P. Léger, P. and S. Sénécal (2021).Smartwatches are more distracting than mobile phones while driving: Results from an experimental study, \emph{Accident Analysis & Prevention}, 149, 1-9.
}
\keyword{datasets}
