% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BSJ92}
\alias{BSJ92}
\title{Baumann, Seifert-Kessell and Jones (1992)}
\format{
A data frame with 66 rows and 6 variables:
\describe{
  \item{\code{group}}{[factor] experimental group, one of directed reading-thinking activity (DRTA), think-aloud (TA) and directed reading group (DR)}
  \item{\code{pretest1}}{[integer] score (out of 16) on pretest for the error detection task}
  \item{\code{pretest2}}{[integer] score (out of 15) on second pretest using a comprehension monitoring questionnaire}
  \item{\code{posttest1}}{[integer] score (out of 16) on an error detection task}
  \item{\code{posttest2}}{[integer] score (out of 18) on an expanded comprehension monitoring questionnaire}
  \item{\code{posttest3}}{[integer] number of correct item, ranging from 0 to 56 on the Degrees of Reading Power test, a cloze test}
}
}
\source{
[carData::Baumann]
}
\usage{
BSJ92
}
\description{
Experimental study on the effectiveness of different reading strategies on understanding of children.
}
\references{
Baumann, J. F., Seifert-Kessell, N., & Jones, L. A. (1992). \emph{Effect of Think-Aloud Instruction on Elementary Students’ Comprehension Monitoring Abilities}. Journal of Reading Behavior, 24(\bold{2}), 143–172. \url{https://doi.org/10.1080/10862969209547770}
}
\keyword{datasets}
