% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{C16}
\alias{C16}
\title{Chandler (2016) replication}
\format{
A data frame with 120 rows and 4 variables:
\describe{
  \item{\code{anchor}}{[factor] anchor, either \code{round} or \code{precise}}
  \item{\code{magnitude}}{[factor] magnitude of anchor, 0 or 1}
  \item{\code{gender}}{[factor] participant gender, either \code{female}, \code{male} or \code{other}}
  \item{\code{madjust}}{[double] mean adjustement}
}
}
\source{
Open Science Foundation,  \url{https://osf.io/aaudl}, licensed under CC0 1.0 Universal
}
\usage{
C16
}
\description{
Replication by Chandler (2016) of Study 4a of Janiszewski and Uy (2008). Both studies measured the amount of adjustment when presented with vague or precise range of value for objects, with potential encouragement for adjusting more the value.
}
\references{
Chandler, J. J. (2016). \emph{Replication of Janiszewski & Uy (2008, Psychological Science, study 4b)}. OSF. \url{https://osf.io/aaudl}

Janiszewski, C., & Uy, D. (2008). \emph{Precision of the anchor influences the amount of adjustment}. Psychological Science, 19(\bold{2}), 121–127. \url{https://doi.org/10.1111/j.1467-9280.2008.02057.x}
}
\keyword{datasets}
