% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{C18}
\alias{C18}
\title{Clayton (2018)}
\format{
A data frame with 101 rows and 7 variables:
\describe{
  \item{\code{age}}{[integer] age of respondant}
  \item{\code{female}}{[integer] indicator variable, 1 if female, 0 otherwise}
  \item{\code{village}}{[factor] village name}
  \item{\code{ED}}{[integer] electoral division code}
  \item{\code{quota}}{[integer] experimental condition, whether a gender quota is enforced (1) or not (0)}
  \item{\code{dscore}}{[double] d-score for Implicit Association Tests (IAT)}
  \item{\code{under25}}{[integer] indicator variable, 1 if the person is 25 or under, 0 otherwise}
}
}
\usage{
C18
}
\description{
Study by A. Clayton in Lesotho "to test whether citizens express less gender bias after 6-years of exposure to quota-elected women village representatives". The data are clustered by village and electoral district.
}
\references{
Clayton, A. (2018). \emph{Do Gender Quotas Really Reduce Bias? Evidence from a Policy Experiment in Southern Africa}. Journal of Experimental Political Science, 5(\bold{3}), 182-194. \url{https://doi.org/10.1017/XPS.2018.8}
}
\keyword{datasets}
