% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{C22}
\alias{C22}
\title{Curley et al. (2022)}
\format{
A data frame with 256 rows and 6 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{anchor}}{[factor] order of evidence, (\code{strong-first} versus \code{weak-first})}
  \item{\code{vignette}}{[factor] homocide vignette}
  \item{\code{verdictsyst}}{[factor] verdict system, either \code{two} or \code{three} possible outcomes.}
  \item{\code{guilt}}{[double] final belief of guilt score, with lower values representing no belief of guilt}
  \item{\code{pjaq}}{[integer] PJAQ score, a 29-item questionnaire, which measures total pre-trial bias}
}
}
\usage{
C22
}
\description{
This study aimed at finding the impact of having multiple options for jurors as is the case in Scotland (guilty, not guilty, not proven), relative to the usual setting of guilty or not guilty verdict.

The response was measured using the sum of the scores for 11 ratings (rounds) on a Likert scale ranging from 0 (no evidence of guilt) to 14 (guilty).
}
\references{
Lee J. Curley, Jennifer Murray, Rory MacLean, James Munro, Martin Lages, Lara A. Frumkin, Phyllis Laybourn & David Brown (2022) \emph{Verdict spotting: investigating the effects of juror bias, evidence anchors and verdict system in jurors}, Psychiatry, Psychology and Law, 29(\bold{3}), 323-344, \url{https://doi.org/10.1080/13218719.2021.1904450}
}
\keyword{datasets}
