% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DA22_E2}
\alias{DA22_E2}
\title{Duke and Amir (2022), Experiment 2}
\format{
A data frame with 325 rows and 7 variables:
\describe{
  \item{\code{gender}}{[factor] gender of respondant, one of \code{female}, \code{male} or \code{other}}
  \item{\code{age}}{[integer] age of respondant}
  \item{\code{format}}{[factor] experimental condition, either \code{quantity-integrated} or \code{quantity-sequential}}
  \item{\code{purchased}}{[integer] binary variable indicating whether the respondant purchased the item (\code{1}) or not (\code{0})}
  \item{\code{amount}}{[integer] amount purchased}
  \item{\code{meanval}}{[double] mean scored value}
  \item{\code{elength}}{[integer] length of self-reported explanation for their choice (number of characters)}
}
}
\source{
Research Box 602, \url{https://researchbox.org/602}, unspecified license.
}
\usage{
DA22_E2
}
\description{
This is a subset of the data from Experiment 2 of Duke and Amir (2022), excluding all answers deemed nonsensical by at least one of the lab assistants. The study investigated the impact of presenting customers with a sequential choice (first decide whether or not to buy, then pick quantity) as opposed to an integrated decision (choose not to buy, or one of different quantities) on sales and the number of items bought.
}
\references{
Duke, K.E. and O. Amir (2022+). \emph{The Importance of Selling Formats: When Integrating Purchase and Quantity Decisions Increases Sales}, Marketing Science, \url{https://doi.org/10.1177/0956797614545886}
}
\keyword{datasets}
