% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DA23_E3}
\alias{DA23_E3}
\title{Duke and Amir (2023), Experiment 3}
\format{
A data frame with 399 rows and 6 variables:
\describe{
  \item{\code{format}}{[factor] experimental condition, either \code{quantity-integrated} or \code{quantity-sequential}}
  \item{\code{purchased}}{[integer] binary variable indicating whether the respondant purchased the item (\code{1}) or not (\code{0})}
  \item{\code{amount}}{[integer] amount of items purchased}
  \item{\code{firstbrand}}{[factor] company name offering the product, counterbalanced}
  \item{\code{gender}}{[factor] gender of respondant, one of \code{female}, \code{male} or \code{other}}
  \item{\code{age}}{[integer] age of respondant}
}
}
\source{
Research Box 602, \url{https://researchbox.org/602}, licensed under CC BY 4.0
}
\usage{
DA23_E3
}
\description{
This is a data from Experiment 3 of Duke and Amir (2022), who studied the impact of sequential vs integrated quantity purchase. The authors manipulated the company name (same ratings, same product and price) and offered participants the option to purchase up to two items on sale from either, with brand order counterbalanced. The study investigated the impact of presenting customers with a sequential choice (first decide whether or not to buy, then pick quantity) as opposed to an integrated decision (choose not to buy, or one of different quantities) on sales and the number of items bought.
}
\examples{
mod <- glm(purchased ~ format * firstbrand, data = DA23_E3)
summary(mod)
}
\references{
Duke, K.E. and O. Amir (2023). \emph{The Importance of Selling Formats: When Integrating Purchase and Quantity Decisions Increases Sales}, Marketing Science,42(1), 87-109. \doi{10.1287/mksc.2022.1364}
}
\keyword{datasets}
