% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DM22_S1_long}
\alias{DM22_S1_long}
\title{Dorison and Minson (2022), Study 1 (long format)}
\format{
A data frame with 734 rows and 4 variables:
\describe{
  \item{\code{anxiety}}{[numeric] anxiety score for average of four nine point Likert scales}
  \item{\code{target}}{[factor] categorical variable, \code{self} or \code{other}}
  \item{\code{vote}}{Who are you most likely to vote for in the upcoming U.S. presidential election (2020)? either Republican candidate (1) or Democratic candidate (2)}
  \item{\code{group}}{[factor] group label}
}
}
\source{
ResearchBox 577, \url{https://researchbox.org/577}, licensed under CC BY 4.0
}
\usage{
DM22_S1_long
}
\description{
Long-format database. Participants (American citizens) in the study were paired with people from the opposing party (Republican or Democrat) and asked to
perform a 10 minutes presentation to discuss beliefs about the best candidate, with pointers to subjects
such as Covid19, environmental policy and immigration. Only people who interacted via the chat are kept.
}
\references{
Dorison CA, Minson JA. (2022). You can’t handle the truth! Conflict counterparts over-estimate each other’s feelings of self-threat. \emph{Organizational Behavior and Human Decision Processes}, 170, 104147. \doi{10.1016/j.obhdp.2022.104147}
}
\keyword{datasets}
