% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DM22_S2}
\alias{DM22_S2}
\title{Dorison and Minson (2022), Study 2}
\format{
A data frame with 319 rows and 5 variables:
\describe{
  \item{\code{age}}{[integer] age of participant}
  \item{\code{gender}}{[factor] gender of participant, either \code{male} or \code{female} or \code{non-binary/other}}
  \item{\code{ideology}}{[factor] self-reported political orientation, with choices very/somewhat/slightly liberal (1/2/3), neither (4), to slightly/somewhat/very conservative (5/6/7).}
  \item{\code{anxiety}}{[numeric] average of the four items from the anxiety scale. 0-8 scale, where higher numbers indicate more anxiety}
  \item{\code{anger}}{[numeric] average of the four items from the anger scale. 0-8 scale, where higher numbers indicate more anger}
}
}
\source{
ResearchBox 577, \url{https://researchbox.org/577}, licensed under CC BY 4.0
}
\usage{
DM22_S2
}
\description{
In Study 2, participants were asked to watch a video of Bernie Sanders talking about immigration, and Ted Cruz talking about tax reform (each for two minutes), and forecast the emotions of a conservative watching the video.
}
\references{
Dorison CA, Minson JA. (2022). You can’t handle the truth! Conflict counterparts over-estimate each other’s feelings of self-threat. \emph{Organizational Behavior and Human Decision Processes}, 170, 104147. \doi{10.1016/j.obhdp.2022.104147}
}
\keyword{datasets}
