% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DM22_S3}
\alias{DM22_S3}
\title{Dorison and Minson (2022), Study 3}
\format{
A data frame with 399 rows and 5 variables:
\describe{
  \item{\code{appraisal}}{[numeric] average of four appraisal items  (9 points scale from 0 to 8, with lower value indicating Not even the slightest bit to more strongly than ever before)}
  \item{\code{threat}}{[numeric] average of two threat items without descriptors  (9 points scale)}
  \item{\code{general}}{[numeric] average of four general items (9 points scale)}
  \item{\code{condition}}{[factor] condition randomly assigned to}
  \item{\code{naiverealism}}{[integer] answer to single-item naive realism question, "to what extent would the arguments your opponent makes lead you to question whether your beliefs on this issue are fundamentally correct and objective?", with not at all (1), to very much (5)}
}
}
\source{
ResearchBox 577, \url{https://researchbox.org/577}, licensed under CC BY 4.0
}
\usage{
DM22_S3
}
\description{
MTurk participants were asked to write a topic for which they held a strong opinion (gay rights, abortion, gun control, immigration, etc.) and imagine a discussion with a person having opposite views from theirs.
}
\references{
Dorison CA, Minson JA. (2022). You can’t handle the truth! Conflict counterparts over-estimate each other’s feelings of self-threat. \emph{Organizational Behavior and Human Decision Processes}, 170, 104147. \doi{10.1016/j.obhdp.2022.104147}
}
\keyword{datasets}
