% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{DM22_S4}
\alias{DM22_S4}
\title{Dorison and Minson (2022), Study 4}
\format{
A data frame with 322 rows and 2 variables:
\describe{
  \item{\code{wager}}{amount wagered on ability to persuade counterpart (0-50 continuous variable)}
  \item{\code{condition}}{between-subject experimental condition, either \code{afraid} or \code{not}}
}
}
\source{
ResearchBox 577, \url{https://researchbox.org/577}, licensed under CC BY 4.0
}
\usage{
DM22_S4
}
\description{
Exploration of the behavioral consequences of overestimating self-threat.
}
\references{
Dorison CA, Minson JA. (2022). You can’t handle the truth! Conflict counterparts over-estimate each other’s feelings of self-threat. \emph{Organizational Behavior and Human Decision Processes}, 170, 104147. \doi{10.1016/j.obhdp.2022.104147}
}
\keyword{datasets}
