% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{FBC66_T1}
\alias{FBC66_T1}
\title{Flavell, Beach and Chinsky (1966), Table 1}
\format{
A data frame with 9 rows and 3 variables:
\describe{
  \item{\code{grade}}{[factor] grade of student, one of kindergarden (\code{K}), second (\code{2}) or fifth (\code{5}) grade}
  \item{\code{frequency}}{[factor] category indicating how often students verbalized during task, one of \code{0}, \code{1-2} or \code{3+}}
  \item{\code{count}}{[integer] count of the number of student}
}
}
\usage{
FBC66_T1
}
\description{
Three by three contingency table for the data presented in Table 1 of Flavell, Beach and Chinsky (2021), in long format. The authors recruited 20 students per grade.
}
\examples{
chisq.test(xtabs(count ~ grade + frequency, data = FBC66_T1))

}
\references{
Flavell, J. H., Beach, D. R., & Chinsky, J. M. (1966). Spontaneous verbal rehearsal in a memory task as a function of age. \emph{Child Development}, 37, 283-299. \doi{10.2307/1126804}
}
\keyword{datasets}
