% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{GK14_S3}
\alias{GK14_S3}
\title{Grossman and Kross (2014) study 3}
\format{
A data frame with 469 rows and 6 variables:
\describe{
  \item{\code{age}}{[factor] age group, either \code{young} or \code{old}}
  \item{\code{condition}}{[factor] experimental condition, one of \code{self immersed}, \code{self distanced}, \code{other immersed} and \code{other distanced}}
  \item{\code{limits}}{[double] score for the "recognition of limits of knowledge"}
  \item{\code{compr}}{[double] score for the "search for a compromise"}
  \item{\code{persp}}{[double] score for the "consideration of others' perspectives"}
  \item{\code{change}}{[double] score for the "recognition of change"}
}
}
\usage{
GK14_S3
}
\description{
This study focuses on the wise reasoning of people when dealing with problems of other (vs their own) depending on their perspective.
}
\references{
Grossmann, I., & Kross, E. (2014). \emph{Exploring Solomon’s Paradox: Self-Distancing Eliminates the Self-Other Asymmetry in Wise Reasoning About Close Relationships in Younger and Older Adults.} Psychological Science, 25(\bold{8}), 1571–1580. \url{https://doi.org/10.1177/0956797614535400}
}
\keyword{datasets}
