% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{GSBE10}
\alias{GSBE10}
\title{Garcia et al. (2010)}
\format{
A data frame with 129 rows and 5 variables:
\describe{
  \item{\code{protest}}{[factor] experimental condition}
  \item{\code{likeability}}{[double] likeability score for Catherine, average of 6 items}
  \item{\code{angry}}{[double] Likert scale for "I feel angry at Catherine"}
  \item{\code{respeval}}{[double] evaluation of the appropriateness of the response of Catherine}
  \item{\code{sexism}}{[double] average of 6 Likert scales for the Modern Sexism Scale assessing pervasiveness of gender discrimination}
}
}
\source{
Hayes, A. F. (2021), Introduction to Mediation, Moderation, and Conditional Process Analysis: A Regression-Based Approach, 3rd edition, Guilford Press.
}
\usage{
GSBE10
}
\description{
These data are from a study on gender discrimination. Participants were put with a file where a women was turned down promotion in favor of male colleague despite her being clearly more experimented and qualified. The authors manipulated the decision of the participant to this decision, either choosing not to challenge the decision (\code{no protest}), a request to reconsider based on individual qualities of the applicants (\code{individual}) and a request to reconsider based on abilities of women (\code{collective}).

All items were measured using scales constructed using items measured using Likert scales ranging from strongly disagree (\code{1}) to strongly agree (\code{7}).
}
\references{
Garcia, D.M., Schmitt, M.T., Branscombe, N.R. and Ellemers, N. (2010), Women's reactions to ingroup members who protest discriminatory treatment: The importance of beliefs about inequality and response appropriateness. \emph{European Journal of Social Psycholog.}, \bold{40}, 733-745. \doi{10.1002/ejsp.644}
}
\keyword{datasets}
