% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{GSC20_E2}
\alias{GSC20_E2}
\title{Gaesser, Shimura and Cikara (2020), Experiment 2}
\format{
A data frame with 5810 rows and 8 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{help}}{[integer] experimental manipulation either episodic helping, or no help (\code{control})}
  \item{\code{group}}{[factor] group, either \code{in} or \code{out} group}
  \item{\code{story}}{[factor] story narrative, one from \code{A} to \code{J}}
  \item{\code{scene_imagery}}{[double] standardized score for scene imagery, "Did you imagine a scene of helping the person in your mind?",(1=not at all, 4=moderate, 7=coherently and clearly)}
  \item{\code{person_imagery}}{[double] standardized score for person imagery, "Did you visualize the person in your mind?" (1=vividly, as if you were currently there, 4=moderate, 7=not at all)}
  \item{\code{persp_taking}}{[double] standardized score on the perspective, "Did you consider the person's thoughts and feelings?" (1=not at all, 4=moderate, 7=strongly considered)}
  \item{\code{willing_help}}{[double] standardized score for willingess to help, "How likely would you be to help in this situation?" (1=not at all, 4=moderately willing, 7=very willing)}
}
}
\usage{
GSC20_E2
}
\description{
Data obtained by running an online survey on Qualtrics. Self-report questions on thoughts, feeling and self-attributes, in the context of a competition between parties. Participants reported their political preferences, but were randomly allocated to either Democrat or Republican team. The description implied a problem-solving challenge (which did not occur).
}
\note{
The participant demographic information is given in the attributes as metadata.
}
\examples{
# Authors coded -1/1 for group and help
# Unweighted regression (different weights depending on sample size)
GSC20_E2$group_w <- ifelse(GSC20_E2$group=="in", 1, -1)
help_num <- -with(GSC20_E2, sum(help == "episodic") / sum(help == "control"))
GSC20_E2$help_w <- ifelse(GSC20_E2$help == "episodic", 1, -help_num)
GSC20_E2$group_bin <- ifelse(GSC20_E2$group == "control", -1, 1)
GSC20_E2$help_bin <- ifelse(GSC20_E2$help == "in", 1, -1)
lme4::lmer(willing_help ~ group*help + (1 + group | id) + (1 | story), data = GSC20_E2)
lme4::lmer(willing_help ~ group_w*help_w + (1 + group_w | id) + (1 | story), data = GSC20_E2)
}
\references{
Gaesser, B., Shimura, Y., & Cikara, M. (2020). Episodic simulation reduces intergroup bias in prosocial intentions and behavior. \emph{Journal of Personality and Social Psychology}, 118(\bold{4}), 683–705. \url{https://doi.org/10.1037/pspi0000194}
}
\keyword{datasets}
