% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{HS22_P}
\alias{HS22_P}
\title{Hart and Schweitzer (2022) pilot study}
\format{
A data frame with 5808 rows and 7 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{score}}{[double] ERRO score}
  \item{\code{item}}{[factor] item }
  \item{\code{category}}{[factor] category of the item, either \code{product} or \code{service}}
  \item{\code{negotcl}}{[factor] took negotiation class}
  \item{\code{gender}}{[factor] gender of participant, either \code{female} or \code{male}}
  \item{\code{age}}{[integer] age of the participant}
}
}
\source{
Research Box 366, \url{https://researchbox.org/366}, licensed under CC BY 4.0
}
\usage{
HS22_P
}
\description{
The authors measured the economic relevance of relational outcomes (ERRO),
by substracting the participants rating for good versus bad relationships,
which rated from -10 (much worse) to 10 (much better). The data are presented in long format for each
of the three items/services split by category.
}
\examples{
# Repeated measure ANOVA
(rmod <- afex::aov_ez(
    id = "id",
    dv = "score",
    within = "item",
    data = HS22_P))
emm <- emmeans::emmeans(object = rmod, specs = "item")
emmeans::contrast(emm, list(service_vs_product = c(1, -1, 1, -1, 1, -1)/3))
}
\references{
Hart E, Schweitzer ME. (2022) \emph{When we should care more about relationships than favorable deal terms in negotiation: The economic relevance of relational outcomes (ERRO)'. Organizational Behavior and Human Decision Processes}. 168:104108.
\doi{10.1016/j.obhdp.2021.104108}
}
\keyword{datasets}
