% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{HS22_S2}
\alias{HS22_S2}
\title{Hart and Schweitzer (2022), Study 2}
\format{
A data frame with 295 rows and 6 variables:
\describe{
  \item{\code{erro}}{[factor] whether the offer is for a \code{service} (cleaning), or for a \code{product} (selling a vacuum cleaner)}
  \item{\code{relation}}{[double] average relationship importance}
  \item{\code{deal}}{[double] average deal importance}
  \item{\code{slope}}{[double] linear slope between choices 1 and 2}
  \item{\code{dealpts1}}{[double] number of points out of four allocated to deal terms for choice 1. The balance is allocated to relation.}
  \item{\code{dealpts2}}{[double] number of points out of seven allocated to deal terms for choice 2. The balance is allocated to relation.}
}
}
\source{
Research Box 366, \url{https://researchbox.org/366}, licensed under CC BY 4.0
}
\usage{
HS22_S2
}
\description{
Between-subject mediation study. The authors investigated the trade-offs between relational and deal terms concerns.
Comparisons are made between high ERRO (for service purchase from someone providing his services as a house cleaner, versus from a seller of a vacuum cleaner).
Participants were asked to allocate 4 points (for choice 1) to either deal or relationship and seven points for choice 2. The \code{slope} variable is obtained
`((7-dealpts2) - (4-dealpts1))/(dealpts2 - dealpts1)`
}
\references{
Hart E, Schweitzer ME. (2022) \emph{When we should care more about relationships than favorable deal terms in negotiation: The economic relevance of relational outcomes (ERRO)'. Organizational Behavior and Human Decision Processes}. 168:104108.
\doi{10.1016/j.obhdp.2021.104108}
}
\keyword{datasets}
