% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{JCD14_S2}
\alias{JCD14_S2}
\title{Johnson, Cheung and Donnellan (2014), Study 2}
\format{
A data frame with 126 rows and 20 variables:
\describe{
  \item{\code{condition}}{[factor] experimental condition, either \code{control} or \code{cleanliness}}
  \item{\code{mvignette}}{[double] mean of vignettes}
  \item{\code{dog}}{[double] Likert scale for the dog vignette}
  \item{\code{trolley}}{[double] Likert scale for the trolley vignette}
  \item{\code{wallet}}{[double] Likert scale for the wallet vignette}
  \item{\code{plane}}{[double] Likert scale for the plane vignette}
  \item{\code{resume}}{[double] Likert scale for the resume vignette}
  \item{\code{kitten}}{[double] Likert scale for the kitten vignette}
  \item{\code{age}}{[integer] age of participant}
  \item{\code{gender}}{[factor] gender of participant, either \code{male} or \code{female}}
}
}
\usage{
JCD14_S2
}
\description{
This dataset is for a replication of a
 study of Schnall, Benton, and Harvey (2008),
 who hypothesied that cleanliness (washing hands)
reduces the severity of moral judgments. The study failed to replicate.
}
\details{
The vignettes (\code{dog}, \code{trolley}, \code{wallet}, \code{plane}, \code{resume}, \code{kitten}) are measured using a Likert scale ranging from nothing wrong at all (\code{1}) to extremely wrong (\code{7}).
}
\references{
Johnson, D.J., F. Cheung and M.B. Donnellan (2014), \emph{Does Cleanliness Influence Moral Judgments}?, Social Psychology, 45(\bold{3}), 209-215, \url{https://doi.org/10.1027/1864-9335/a000186}
}
\keyword{datasets}
