% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{KSBN24_S3}
\alias{KSBN24_S3}
\title{Koetke et al. (2024), Study 3}
\format{
A data frame with 369 rows and 10 variables:
\describe{
  \item{\code{ih}}{[double] CIHS score for intellectual humility}
  \item{\code{condition}}{[factor] experimental manipulation, either \code{low}, \code{high} or \code{neutral}}
  \item{\code{gender}}{[factor] gender of participant, either \code{male}, \code{female} or \code{other}}
  \item{\code{gender}}{[factor] experimental manipulation for gender of scientist, either \code{male} (Robert) or \code{female} (Sandra).}
  \item{\code{METI}}{[double] METI scale for the trust in science}
  \item{\code{expertise}}{[double] METI subscale for expertise}
  \item{\code{integrity}}{[double] METI subscale for integrity}
  \item{\code{benevolence}}{[double] METI subscale for benevolence}
  \item{\code{research}}{[double] score for support for the research}
  \item{\code{intention}}{[double] score for intention to follow research models}
}
}
\source{
Open Science Foundation, \url{https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961}, unspecified license
}
\usage{
KSBN24_S3
}
\description{
Data for a two-way between-subject ANOVA as a function of \code{condition} and \code{gender}.
}
\references{
Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. \emph{Nature Human Behaviour} (2024). \doi{10.1038/s41562-024-02060-x}. Preprint at \url{https://osf.io/preprints/psyarxiv/ah3fp/}
}
\keyword{datasets}
