% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{KSBN24_S4}
\alias{KSBN24_S4}
\title{Koetke et al. (2024), Study 4}
\format{
A data frame with 371 rows and 10 variables:
\describe{
  \item{\code{METI}}{[double] METI scale for the trust in the scientist}
  \item{\code{research}}{[double] support for Dr. Wilson's research}
  \item{\code{ih}}{[double] perceived intellectual humility score}
  \item{\code{competence}}{[double] score for the question "As viewed by society, how competent is Dr."}
  \item{\code{warmth}}{[double] score for the question "As viewed by society, how warm is Dr."}
  \item{\code{condition}}{[factor] experimental factor for intellectual humility, either \code{low}, \code{high} or \code{neutral}}
  \item{\code{race}}{[factor] experimental factor for race, one of \code{latinx}, \code{black} (African American) or \code{white}}
  \item{\code{politideo}}{[factor] political ideology on Prolifics platform, either \code{liberal} or \code{conservative}}
  \item{\code{age}}{[integer] age of participant}
  \item{\code{gender}}{[factor] gender of participant, either \code{male}, \code{female} or \code{other}}
}
}
\source{
Open Science Foundation, \url{https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961}, unspecified license
}
\usage{
KSBN24_S4
}
\description{
Two (humility) by three race) between subject design to study participant perception of scientists depending on the intellectual humility.
The authors performed linear mediation analysis of \code{METI} by \code{ih}, \code{research} by \code{METI} and by \code{ih}, separately. They also considered two-way ANOVA models with controls for \code{competence} and \code{warmth} as control variates.
}
\references{
Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. \emph{Nature Human Behaviour} (2024). \doi{10.1038/s41562-024-02060-x}. Preprint at \url{https://osf.io/preprints/psyarxiv/ah3fp/}
}
\keyword{datasets}
