% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{KSBN24_S5}
\alias{KSBN24_S5}
\title{Koetke et al. (2024), Study 5}
\format{
A data frame with 679 rows and 13 variables:
\describe{
  \item{\code{ih}}{[double] CIHS score for intellectual humility of scientist}
  \item{\code{condition}}{[factor] experimental condition for intellectual humility, either \code{low} or \code{high}}
  \item{\code{indep}}{[double] CIHS subscale score for independent of science}
  \item{\code{respect}}{[double] CIHS subscale score for respect of science}
  \item{\code{openness}}{[double] CIHS subscale score for openness of science}
  \item{\code{lackover}}{[double] CIHS subscale score for lack over}
  \item{\code{METI}}{[double] METI scale for the trust in science}
  \item{\code{expertise}}{[double] METI subscale for expertise}
  \item{\code{integrity}}{[double] METI subscale for integrity}
  \item{\code{benevolence}}{[double] METI subscale for benevolence}
  \item{\code{research}}{[double] score for support for the research}
  \item{\code{warmth}}{[double] score for warmth}
  \item{\code{competence}}{[double] score for competence}
}
}
\source{
Open Science Foundation, \url{https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961}, unspecified license
}
\usage{
KSBN24_S5
}
\description{
Authors perform one-way ANOVA for CIHS score and the subscales as a function of \code{condition}, and the same for the \code{METI} scale and subscales. Three independent linear mediation are tested for research average and METI, with the experimental \code{condition} mediated by the intellectual humility score \code{ih}, and \code{research} average by \code{condition} also mediated by \code{METI}.
}
\references{
Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. \emph{Nature Human Behaviour} (2024). \doi{10.1038/s41562-024-02060-x}. Preprint at \url{https://osf.io/preprints/psyarxiv/ah3fp/}
}
\keyword{datasets}
