% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{KSBN24_S5L}
\alias{KSBN24_S5L}
\title{Koetke et al. (2024), Study 5 (long format)}
\format{
A data frame with 4074 rows and 5 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{statement}}{[factor] statement of the question}
  \item{\code{pih}}{[double] perceived intellectual humility for the statement}
  \item{\code{trustdr}}{[double] score for trust in doctor Wilson}
  \item{\code{trustresearch}}{[double] score for trust in research results of doctor Wilson}
}
}
\source{
Open Science Foundation, \url{https://osf.io/d3xua/?view_only=0ebe5366331f497da7c78a7c874c2961}, unspecified license
}
\usage{
KSBN24_S5L
}
\description{
Data from Study 5 in long format for the questions related to trust. The authors performed repeated measure ANOVA for each of the three response in term. Also suitable for multivariate analysis of variance. All responses are measured using a Likert scale ranging from 1 (strongly disagree) to 7 (strongly agree).
}
\references{
Koetke, J., Schumann, K., Bowes, S.M. et al. The effect of seeing scientists as intellectually humble on trust in scientists and their research. \emph{Nature Human Behaviour} (2024). \doi{10.1038/s41562-024-02060-x}. Preprint at \url{https://osf.io/preprints/psyarxiv/ah3fp/}
}
\keyword{datasets}
