% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LBJ17_S1A}
\alias{LBJ17_S1A}
\title{Lages, Boyle and Jenkins (2017) Study 1A}
\format{
A data frame with 1760 rows and 6 variables:
\describe{
  \item{\code{nerror}}{[double] number of incorrectly identified letters out of 4}
  \item{\code{adaptation}}{[factor] either preadaptation phase (\code{preadaptation}) or adaptation to contracting motion (\code{contracting})}
  \item{\code{fontsize}}{[factor] font size, in logMAR, one of \code{-0.3}, \code{-0.2}, \code{-0.1} and \code{0}}
  \item{\code{position}}{[factor] position of the letter from left (\code{1}) to right (\code{5})}
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{acuity}}{[factor] acuity group, either \code{normal} or \code{high}}
}
}
\source{
Open Science Foundation, \url{https://osf.io/nav9h/}, unspecified license
}
\usage{
LBJ17_S1A
}
\description{
Data for a study on visual acuity of
participants. The data represent the
number of words correctly detected at different font
size; interest is in effect of illusory contraction on
detection. The participants were dichotomized in two
groups based on their visual acuity based on median split.
}
\examples{
\dontrun{
# Code to reproduce Figure 2a
library(dplyr)
library(ggplot2)
LBJ17_S1A |>
group_by(acuity, fontsize, adaptation, id) |>
 summarize(ncorrect = 20 - sum(nerror))|>
 group_by(acuity, fontsize, adaptation) |>
 summarize(mean = mean(ncorrect))|>
 ggplot(mapping = aes(y = mean,
                      x = as.numeric(fontsize),
                      color = adaptation,
                      shape = acuity)) +
 geom_point() +
 geom_line() +
 labs(x = "Letter Size (logMAR)",
      y = "Number of Letters (Correct)")
}
}
\references{
Lages, M., Boyle, S. C., & Jenkins, R. (2017). \emph{Illusory Increases in Font Size Improve Letter Recognition}. Psychological Science, 28(\bold{8}), 1180–1188. \url{https://doi.org/10.1177/0956797617705391}
}
\keyword{datasets}
