% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LC19_T2}
\alias{LC19_T2}
\title{Lee and Choi (2019), Table 2}
\format{
A data frame with 12 rows and 4 variables:
\describe{
  \item{\code{text}}{[factor] number of toothbrush in text description}
  \item{\code{image}}{[factor] number of toothbrushes displayed in image}
  \item{\code{expected}}{[factor] expected number of toothbrushes}
  \item{\code{count}}{[integer] count of occurence}
}
}
\source{
Mendeley, \url{https://doi.org/10.17632/r9t7hh7cy3.1}, licensed under CC BY 4.0
}
\usage{
LC19_T2
}
\description{
Contingency table for Study 3, measuring the expected number of products as a function of the text description and image showing either one or six toothbrushes.
}
\details{
Distributed under CC-BY 4.0 license
}
\references{
K. Lee and J. Choi (2019). \emph{Image-text inconsistency effect on product evaluation in online retailing}, Journal of Retailing and Consumer Services, 49, 279-288, \url{https://doi.org/10.1016/j.jretconser.2019.03.015}
}
\keyword{datasets}
