% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LHC26_S1}
\alias{LHC26_S1}
\title{Liu et al. (2026), Study 1}
\format{
A data frame with 207 rows and 6 variables:
\describe{
  \item{\code{choice}}{[factor] response of participant, either artificial intelligence (\code{AI}) or \code{human}}
  \item{\code{strength}}{[double] average Z score of scales capturing the job application strength}
  \item{\code{completionrate}}{[double] job completion rate}
  \item{\code{jobcompleted}}{[double] number of job completed}
  \item{\code{hours}}{[double] total work hours}
  \item{\code{earnings}}{[double] total earnings}
}
}
\usage{
LHC26_S1
}
\description{
Authors consider online applications with different strengths (taken to be the average of Z scores for platform-provided metrics, which are: completion rate, number of job completed, work hours, and past earnings). The choice of the participant is whether to rely on an AI or human evaluation for the job application.
}
\references{
Liu, Q., Häubl, G. and N. Castelo (2026). Consumers with Weaker Applications Are Less Receptive to Algorithmic Evaluation, \emph{Journal of Consumer Research}, \doi{10.1093/jcr/ucag033}
}
\keyword{datasets}
