% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LJLSFBM20}
\alias{LJLSFBM20}
\title{Labonte-LeMoyne et al. (2020)}
\format{
A data frame with 296 rows and 14 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{order}}{[integer] order in which tasks were performed, termed "fatigue" in the paper}
  \item{\code{position}}{[factor] work position, either \code{standing} or \code{sitting}}
  \item{\code{phys_demand}}{[factor] manipulation using either \code{mouse} or \code{touchpad}}
  \item{\code{task_diff}}{[factor] task difficulty, either \code{easy} or \code{difficult}}
  \item{\code{ies}}{[double] inverse efficiency score global stimulus}
  \item{\code{central_alpha}}{[double] alpha for central region}
  \item{\code{parietal_alpha}}{[double] alpha for parietal region}
  \item{\code{central_beta}}{[double] beta for central region}
  \item{\code{parietal_beta}}{[double] alpha for parietal region}
  \item{\code{bmi}}{[double] body mass index}
  \item{\code{sex}}{[factor] sex, either \code{man} or \code{woman}}
  \item{\code{attention}}{[double] attention scale, average of 6 items}
  \item{\code{satisfaction}}{[double] satisfaction score, a scale composed of 3 items}
}
}
\source{
Shang Lin Chen, personal communication, distributed under CC BY-NC-SA 4.0
}
\usage{
LJLSFBM20
}
\description{
Perceived performance for standing desk (relative to sitting desk) measured from questionnaires and obtained from electroencephalograph (EEG) data.
}
\references{
Labonté-LeMoyne, E., Jutras, M.-A., Léger, P.-M., Sénécal, S., Fredette, M., Begon, M., and Mathieu, M.-E. (2020). \emph{Does Reducing Sedentarity With Standing Desks Hinder Cognitive Performance? Human Factors}, 62(\bold{4}), 603–612. \doi{10.1177/0018720819879310}
}
\keyword{datasets}
