% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LKUK24_S3}
\alias{LKUK24_S3}
\title{Lin et al. (2024), Study 3}
\format{
A data frame with 804 rows and 8 variables:
\describe{
  \item{\code{appropriation}}{[double] appropriation scale built from five items: higher values indicate increased perception of cultural appropriation}
  \item{\code{purchase}}{[double] intention to purchase; lower scores indicate lower values}
  \item{\code{attitude}}{[double] attitude towards brand: lower values indicate less confidence}
  \item{\code{politideo}}{[factor] political ideology, either \code{conservative} or \code{liberal}}
  \item{\code{prodcat}}{[factor] product category, either \code{food} or \code{apparel}}
  \item{\code{histpowimbal}}{[factor] historical power imbalance, either \code{control} if no mention is made or \code{primed} if historical context is provided to participants.}
  \item{\code{age}}{[integer] age of the participant}
  \item{\code{gender}}{[factor] gender of the participant, male, female, other/neither or non-response}
}
}
\source{
Research Box 1401, \url{https://researchbox.org/1401}, licensed under CC BY 4.0
}
\usage{
LKUK24_S3
}
\description{
Study 3 focused on mediation by moderation, looking at the impact of historical power imbalance
in perceptions of cultural appropriation. Participants were exposed about the story of an American department store which hosts annual parties for customers.
Based on a pilot study, product categories (apparel and beauty)
were mentioned more frequently (71%) than was food and cuisine (15%). The purpose of the study is to compare,
 across liberal and conservative (balanced) the impact of emphasizing the historical power imbalance.
 The study results in a 2 by 2 by 2 three-way between-subject ANOVA.
}
\details{
The response variables are based on a factor analysis of different items on the questionnaire based on Likert scales ranging from 1 to 7.
}
\references{
J.D. Lin, N.Y.J. Kim, E. Uduehi and A. Keinan (2024). \emph{Culture for Sale: Unpacking Consumer Perceptions of Cultural Appropriation}, Journal of Consumer Research, 51(3), 571–594. \doi{10.1093/jcr/ucad076}
}
\keyword{datasets}
