% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LKUK24_S4}
\alias{LKUK24_S4}
\title{Lin et al. (2024), Study 4}
\format{
A data frame with 711 rows and 11 variables:
\describe{
  \item{\code{appropriation}}{[double] perception of cultural appropriation}
  \item{\code{appreciation}}{[double] }
  \item{\code{intent}}{[double] brand intentions, whether brands actions are considered cultural appropriation?}
  \item{\code{purchase}}{[double] purchase interest for the cookbook}
  \item{\code{attitude}}{[double] brand attitude}
  \item{\code{politideo}}{[factor] political ideology, either \code{conservative} or \code{liberal}}
  \item{\code{chefdax}}{[factor] whether Chef Dax was \code{black} or \code{not black}}
  \item{\code{brandaction}}{[factor] description of the way recipes were obtained by Chef Dax, either \code{peeking}, with \code{permission} or no mention (\code{control}).}
  \item{\code{age}}{[integer] age of the participant}
  \item{\code{gender}}{[factor] gender of the participant, male, female, other/neither or non-response}
  \item{\code{minority}}{[factor] does participant identify as a member of a racial minority? Either \code{yes} or \code{no}}
}
}
\source{
Research Box 1401, \url{https://researchbox.org/1401}, licensed under CC BY 4.0
}
\usage{
LKUK24_S4
}
\description{
Study 4 focused on cultural appropriation for soul food recipe cookbook from Chef Dax, who was
either black (or not), manipulating the description of the way he obtained the recipes (by peeking without permission in kitchens, by asking permission or no mention for control).
Authors postulated that the perception of appropriation would vary by political ideology (liberal or conservative).
The study results in a 3 by 2 by 2 three-way between-subject ANOVA, followed by moderated mediation.
}
\details{
The Qualtric file contains the questions. Some variables seem to be aggregates of various Likert scales, but there are no associated questions.
}
\references{
J.D. Lin, N.Y.J. Kim, E. Uduehi and A. Keinan (2024). \emph{Culture for Sale: Unpacking Consumer Perceptions of Cultural Appropriation}, Journal of Consumer Research, 51(3), 571–594. \doi{10.1093/jcr/ucad076}
}
\keyword{datasets}
