% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LRMM22_S1}
\alias{LRMM22_S1}
\title{Liu et al. (2022+) Study 1}
\format{
A data frame with 200 rows and 5 variables:
\describe{
  \item{\code{appreciation}}{[double] response variable, average Likert scale for appreciation}
  \item{\code{role}}{[factor] experimental factor, either \code{initiator} or \code{responder}}
  \item{\code{form}}{[factor] form of communication, one of \code{email}, \code{text}, \code{phone} or \code{other}}
  \item{\code{gender}}{[factor] gender, one of \code{male}, \code{female} or \code{other}}
  \item{\code{age}}{[double] age of participant}
}
}
\source{
Research Box 289, \url{https://researchbox.org/289}, distributed under a CC BY 4.0 license
}
\usage{
LRMM22_S1
}
\description{
Liu et al. study social interactions and the impact of surprise on people reaching out if this contact is unexpected. Experiment 1 focuses on questionnaires where the experimental condition is the perceived appreciation of reaching out to someone (vs being reached to). The study used a questionnaire administered to 200 American adults recruited on the Prolific Academic platform. The response index consists of the average of four questions measured on a Likert scale ranging from 1 (not at all) to 7 (to a great extent): higher values thus indicate higher appreciation.
}
\references{
Liu, P. J., Rim, S., Min, L., & Min, K. E. (2022+). \emph{The surprise of reaching out: Appreciated more than we think.} Journal of Personality and Social Psychology.  \url{https://doi.org/10.1037/pspi0000402}
}
\keyword{datasets}
