% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LRMM22_S3}
\alias{LRMM22_S3}
\title{Liu et al. (2021), Study 3}
\format{
A data frame with 67 rows and 6 variables:
\describe{
  \item{\code{apprec_resp}}{[integer] appreciation of respondant}
  \item{\code{apprec_init}}{[integer] appreciation of initiator}
  \item{\code{age_resp}}{[double] age of respondant}
  \item{\code{gender_resp}}{[factor] gender of respondant, one of \code{male}, \code{female} or \code{other}}
  \item{\code{age_init}}{[double] age of initiator}
  \item{\code{gender_init}}{[factor] gender of initiator, one of \code{male}, \code{female} or \code{other}}
}
}
\source{
Research Box 289, \url{https://researchbox.org/289}, unspecified license
}
\usage{
LRMM22_S3
}
\description{
Paired data from dyads: participants were asked to "name a fellow college student in their social circle with whom they typically have pleasant encounters but have not interacted (either virtually or in-person)". The experimenter measured the appreciation score via a Likert scale with not at all (\code{1}), slightly (\code{2}), moderately (\code{3}), very \code{4} and extremely (\code{5}). The expectation of the authors was that initiators sending the gift/message would underestimate how much this attention would be valued by respondant
}
\references{
Liu, P. J., Rim, S., Min, L., & Min, K. E. (2022+). \emph{The surprise of reaching out: Appreciated more than we think.} Journal of Personality and Social Psychology.  \url{https://doi.org/10.1037/pspi0000402}
}
\keyword{datasets}
