% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LRMM22_S5b}
\alias{LRMM22_S5b}
\title{Liu et al. (2019), Study 5b}
\format{
A data frame with 402 rows and 5 variables:
\describe{
  \item{\code{appreciation}}{[double] average score questions related to appreciation (appreciate, feel thankful, feel grateful, feel pleased)}
  \item{\code{surprise}}{[double] average score for questions related to unexpectedness and surprise of reaching out}
  \item{\code{condition}}{[factor] experimental condition, either \code{self} or \code{other}}
  \item{\code{age}}{[double] age of participant}
  \item{\code{gender}}{[factor] gender of participant, one of \code{male}, \code{female} or \code{other}}
}
}
\usage{
LRMM22_S5b
}
\description{
The authors postulated that the
underestimation of the appreciation of initiators relative to recipients
was mediated by the degree of surprise of the recipient.
}
\references{
Liu, P. J., Rim, S., Min, L., & Min, K. E. (2022+). \emph{The surprise of reaching out: Appreciated more than we think.} Journal of Personality and Social Psychology.  \url{https://doi.org/10.1037/pspi0000402}
}
\keyword{datasets}
