% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{LWSH23_S3}
\alias{LWSH23_S3}
\title{Leckfor et al (2023), Study 3}
\format{
A data frame with 545 rows and 7 variables:
\describe{
  \item{\code{cond}}{[factor] condition, one of \code{included} (control), \code{ghosted} and \code{directly rejected} and postulated mediator}
  \item{\code{needsatis}}{[double] need for satisfaction, response variable}
  \item{\code{needclosure}}{[double] need for closure, an average of 15 Likert scale, explanatory variable}
  \item{\code{age}}{[integer] age of respondant}
  \item{\code{gender}}{[factor] gender of respondent, either \code{woman}, \code{man} or \code{other}}
  \item{\code{reltype}}{[factor] type of romantic relationship, either \code{casually dating}, \code{friendship}, \code{romantic relationship}}
  \item{\code{rel}}{[double] answer to the questions "how you felt toward the other person before the situation you just described ...", with high values demanding more involvment and commitment to the relationship}
}
}
\usage{
LWSH23_S3
}
\description{
Moderation analysis for need for satisfaction. The author postulated that the type of rejection (control, being ghosted or rejected) would moderate the effect of the need for closure. The authors were postulated "lower needs satisfaction after being ghosted"
}
\examples{
mod <- lm(data = LWSH23_S3, needsatis ~ needclosure * cond)
anova(mod)
# process(data = LWSH23_S3, y = "needsatis", x = "needclosure",
#  w = "cond", mcw = 1, model = 1, plot = 1, intprobe = 1, moments = 1)
}
\references{
Leckfor CM, Wood NR, Slatcher RB, Hales AH (2023) From close to ghost: Examining the relationship between the need for closure, intentions to ghost, and reactions to being ghosted. \emph{Journal of Social and Personal Relationships}, 40(\bold{8}), 2422-2444. \doi{10.1177/02654075221149955}
}
\keyword{datasets}
