% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{MANY18_S18}
\alias{MANY18_S18}
\title{ManyLab2 replication of Risen and Gilovich (2008)}
\format{
A data frame with 4433 rows and 4 variables:
\describe{
  \item{\code{lab}}{[factor] laboratory identifier}
  \item{\code{likelihood}}{[integer] likelihood of bad outcome, from 1 to 10}
  \item{\code{gender}}{[factor] gender of participant, either \code{male} or \code{female}}
  \item{\code{condition}}{[factor] experimental condition, either \code{prepared} or \code{unprepared}}
}
}
\source{
Open Science Foundation, \url{https://osf.io/sg3su}, licensed under CC0 1.0 Universal
}
\usage{
MANY18_S18
}
\description{
The data is a replication with undergraduate students of a study. The authors measured the likelihood of bad outcome when tempting fate by not preparing in advance for a task. The replication data are unbalanced and the study failed to replicate
}
\references{
Klein, R. A., et al. (2018). \emph{Many labs 2: Investigating variation in replicability across samples and settings.} Advances in Methods and Practices in Psychological Science, 1(\bold{4}), 443–490. \url{https://doi.org/10.1177/2515245918810225}
Risen, J. L., and T. Gilovich, (2008). \emph{Why people are reluctant to tempt fate.} Journal of Personality and Social Psychology, 95(\bold{2}), 293–307. \url{https://doi.org/10.1037/0022-3514.95.2.293}
}
\keyword{datasets}
