% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{MP14_S1}
\alias{MP14_S1}
\title{Maglio and Polman (2014), Study 2}
\format{
A data frame with 202 rows and 3 variables:
\describe{
  \item{\code{station}}{[factor] subway station, one of \code{Spadina}, \code{St. George}, \code{Bloor-Yonge} and \code{Sherbourne}}
  \item{\code{direction}}{[factor] direction of travel, either \code{east} or \code{west}}
  \item{\code{distance}}{[integer] subjective distance, a Likert scale ranging from very close (\code{1}) to very far (\code{7})}
}
}
\source{
Open Science Foundation, \url{https://osf.io/7rajd/}, unspecified license
}
\usage{
MP14_S1
}
\description{
Authors measured the subjective distance on travel from one subway station (Bay) and asked the subjective distance to another station, at most two stations away from the location of the experiment. The other factor manipulated was the direction of travel. The data are suitable for a two-way analysis of variance.
}
\details{
The specific question asked to commuters was "How far away does the [name] station feel to you?"
}
\references{
Maglio, S. J., & Polman, E. (2014). \emph{Spatial orientation shrinks and expands psychological distance.} Psychological Science, 25(\bold{7}), 1345–1352. \url{https://doi.org/10.1177/0956797614530571}
}
\keyword{datasets}
