% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{MULTI21_D1}
\alias{MULTI21_D1}
\title{Elliot et al. (2021), dataset 1}
\format{
A data frame with 204 rows and 4 variables:
\describe{
  \item{\code{lab}}{[factor] location of laboratory}
  \item{\code{age}}{[factor] age group, one of \code{5yo}, \code{6yo}, \code{7yo} or \code{10yo}}
  \item{\code{frequency}}{[factor] how often students verbalized during task, one of \code{never}, \code{sometimes} or \code{usually}}
  \item{\code{count}}{[integer]count of the number of children}
}
}
\source{
Open Science Foundation, \url{https://osf.io/vgxkf}, licensed under CC-By Attribution 4.0 International
}
\usage{
MULTI21_D1
}
\description{
Data for Table 2 of Elliot et al. (2021), split by replicating lab. The data are counts of verbalization, suitable for a generalized linear mixed model and chi-square test for contingency data.
}
\references{
Elliott, E. M., C. C. Morey, A.M. AuBuchon et al. (2021). \emph{Multilab Direct Replication of Flavell, Beach, and Chinsky (1966): Spontaneous Verbal Rehearsal in a Memory Task as a Function of Age}. Advances in Methods and Practices in Psychological Science. \url{https://doi.org/10.1177/25152459211018187}
}
\keyword{datasets}
