% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{MULTI21_D2}
\alias{MULTI21_D2}
\title{Elliot et al. (2021), dataset 2}
\format{
A data frame with 2931 rows and 7 variables:
\describe{
  \item{\code{id}}{[factor] unique participant ID}
  \item{\code{lab}}{[factor] location of laboratory}
  \item{\code{taskorder}}{[integer] order of task}
  \item{\code{timing}}{[factor] timing of the task for recall, either \code{delayed} or \code{immediate} or \code{point-and-name}}
  \item{\code{age}}{[factor] age group, one of \code{5yo}, \code{6yo}, \code{7yo} or \code{10yo}}
  \item{\code{mcorrect}}{[integer] number of words correctly recalled, from 1 to 5}
  \item{\code{frequency}}{[factor] how often students verbalized during task, one of \code{never}, \code{sometimes} or \code{usually}}
}
}
\source{
Open Science Foundation, \url{https://osf.io/vgxkf}, licensed under CC-BY Attribution 4.0 International
}
\usage{
MULTI21_D2
}
\description{
Data for a linear mixed model and a two-way analysis of variance with \code{taskorder} and \code{timing}.
}
\references{
Elliott, E. M., C. C. Morey, A.M. AuBuchon et al. (2021). \emph{Multilab Direct Replication of Flavell, Beach, and Chinsky (1966): Spontaneous Verbal Rehearsal in a Memory Task as a Function of Age}. Advances in Methods and Practices in Psychological Science. \doi{10.1177/25152459211018187}
}
\keyword{datasets}
