% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{MV23_S1}
\alias{MV23_S1}
\title{Moon and vanEpps (2022+), Study 1}
\format{
A data frame with 869 rows and 4 variables:
\describe{
  \item{\code{before}}{[integer] did people donate before to this charity? 0 for no, 1 for yes.}
  \item{\code{donate}}{[integer] did people choose to donate to the charity? 0 for no, 1 for yes.}
  \item{\code{condition}}{[factor] experimental condition, either an \code{open-ended} amount or a suggested \code{quantity}}
  \item{\code{amount}}{[double] amount of proposed donation, \code{NA} if the person declined to donate}
}
}
\source{
Research Box 54, \url{https://researchbox.org/54}, licensed under CC BY 4.0
}
\usage{
MV23_S1
}
\description{
Study 1 considers proportion of donators to a charity. Participants in the online panel were provided with an opportunity to win 25$ and donate part of this amount to a charity of their choosing. The data provided include only people who did not exceed this amount and indicated donating a non-zero amount.
}
\examples{
data(MV23_S1, package = "hecedsm")
ctab <- with(MV23_S1, table(donate, condition))
# Chi-square test of independence
chisq_test <- chisq.test(ctab, correct = FALSE)
chisq_test
# Effect size
sqrt(chisq_test$statistic/nrow(MV23_S1))
# Welch t-test
t.test(amount ~ condition, MV23_S1)
}
\references{
Moon, A. and EM VanEpps (2023). \emph{Giving Suggestions: Using Quantity Requests to Increase Donations}, Journal of Consumer Research, 50(1), 190–210. \doi{10.1093/jcr/ucac047}
}
\keyword{datasets}
