% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{PCSCFL24_S1}
\alias{PCSCFL24_S1}
\title{Poirier et al. (2024), Study 1}
\format{
A data frame with 2136 rows and 9 variables:
\describe{
  \item{\code{id}}{[factor] identifier of participant}
  \item{\code{gender}}{[factor] gender of participant, either \code{male}, \code{female} or \code{other}}
  \item{\code{age}}{[integer] age of participant}
  \item{\code{picture}}{[factor] identifier of photo}
  \item{\code{socialpres}}{[factor] experimental factor for social presence, one of \code{no}, \code{medium} or \code{high}}
  \item{\code{valence}}{[integer] positive emotional reaction score}
  \item{\code{purchaseintent}}{[integer] intention to purchase}
  \item{\code{arousal}}{[integer] arousal score}
  \item{\code{diagnosticity}}{[integer] perceived photo diagnosticity}
}
}
\source{
Sara-Maude Poirier, personal communication, distributed under CC BY-NC-SA 4.0
}
\usage{
PCSCFL24_S1
}
\description{
Data for an incomplete one-way ANOVA repeated measure design. The observations are unbalanced per participant and by experimental condition. Participants with missing data for one of the condition were removed from the analysis.
}
\references{
Poirier, S.-M., Cosby, S., Sénécal, S., Coursaris, C. K., Fredette, M. et Léger, P.-M. (2024). \emph{The impact of social presence cues in social media product photos on consumers’ purchase intentions}. Journal of Business Research, 185, 114932. \doi{10.1016/j.jbusres.2024.114932}.
}
\keyword{datasets}
