% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{PCSCFL24_S2}
\alias{PCSCFL24_S2}
\title{Poirier et al. (2024), Study 2}
\format{
A data frame with 587 rows and 9 variables:
\describe{
  \item{\code{staging}}{[factor] experimental factor for staging, termed consumption background (either \code{absent} or \code{present})}
  \item{\code{socialpres}}{[factor] experimental factor for social presence, one of \code{no}, \code{medium} or \code{high}}
  \item{\code{arousal}}{[integer] arousal score}
  \item{\code{valence}}{[integer] positive emotional reaction score}
  \item{\code{gender}}{[factor] gender of participant, either \code{male}, \code{female} or \code{other}}
  \item{\code{age}}{[integer] age of participant}
  \item{\code{diagnosticity}}{[integer] perceived photo diagnosticity}
  \item{\code{purchaseintent}}{double COLUMN_DESCRIPTION}
  \item{\code{warmth}}{integer COLUMN_DESCRIPTION}
}
}
\source{
Sara-Maude Poirier, personal communication, distributed under CC BY-NC-SA 4.0
}
\usage{
PCSCFL24_S2
}
\description{
Data for a between-subject two-way ANCOVA (3x2, social presence vs staging) with gender as covariate for the control, as the high social presence condition showed the back of a woman running on a treadmill.
}
\references{
Poirier, S.-M., Cosby, S., Sénécal, S., Coursaris, C. K., Fredette, M. et Léger, P.-M. (2024). \emph{The impact of social presence cues in social media product photos on consumers’ purchase intentions}. Journal of Business Research, 185, 114932. \doi{10.1016/j.jbusres.2024.114932}.
}
\keyword{datasets}
