% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{RH19_S1}
\alias{RH19_S1}
\title{Roczniewska and Higgins (2019), Study 1}
\format{
A data frame with 198 rows and 12 variables:
\describe{
  \item{\code{change}}{[factor] change outcome depending on school type, either middle school (\code{negative}) or primary school (\code{positive})}
  \item{\code{regfocus}}{[factor] situational regulatory focus manipulation: education change message framing, either \code{prevention} or \code{promotion} message framing}
  \item{\code{fluency}}{[double] mean for fluency items}
  \item{\code{engagement}}{[double] mean for engagement items}
  \item{\code{dprocessjustice}}{[double] mean for change process fairness perceptions items}
  \item{\code{dopenness}}{[double] mean for change openness items}
  \item{\code{promem}}{[double] mean for work regulatory focus promotion items}
  \item{\code{prevem}}{[double] mean for work regulatory focus prevention items}
<<<<<<< HEAD
=======
  \item{\code{gender}}{[factor] gender of teacher, either \code{man} or \code{woman}}
  \item{\code{age}}{[integer] age of teacher}
  \item{\code{tenure}}{[integer] number of years teaching}
  \item{\code{hours_week}}{[integer] weekly number of hours working}
}

A data frame with 198 rows and 12 variables:
\describe{
  \item{\code{change}}{[factor] change outcome depending on school type, either middle school (\code{negative}) or primary school (\code{positive})}
  \item{\code{regfocus}}{[factor] situational regulatory focus manipulation: education change message framing, either \code{prevention} or \code{promotion} message framing}
  \item{\code{fluency}}{[double] mean for fluency items}
  \item{\code{engagement}}{[double] mean for engagement items}
  \item{\code{dprocessjustice}}{[double] mean for change process fairness perceptions items}
  \item{\code{dopenness}}{[double] mean for change openness items}
  \item{\code{promem}}{[double] mean for work regulatory focus promotion items}
  \item{\code{prevem}}{[double] mean for work regulatory focus prevention items}
>>>>>>> update2
  \item{\code{gender}}{[factor] gender of teacher, either \code{male} or \code{female}}
  \item{\code{age}}{[integer] age of teacher}
  \item{\code{tenure}}{[integer] number of years teaching}
  \item{\code{hours_week}}{[integer] weekly number of hours working}
}
}
\usage{
RH19_S1
<<<<<<< HEAD
}
\description{
Study of mediator following changes in the Polish education system, whereby middle school was due to disappear and be replaced with a new system with 8 years of elementary school and four years of high school. The authors surveyed teachers in middle and primary schools with two messages, either prevention or promotion.
}
\references{
M. Roczniewska and Higgins, E.T. (2019). \emph{Messaging organizational change: How regulatory fit relates to openness to change through fairness perceptions}, Journal of Experimental Social Psychology, 85, \url{https://doi.org/10.1016/j.jesp.2019.103882}
=======

RH19_S1
}
\description{
Study of mediator following changes in the Polish education system, whereby middle school was due to disappear and be replaced with a new system with 8 years of elementary school and four years of high school. The authors surveyed teachers in middle and primary schools with two messages, either prevention or promotion.

Study of mediator following changes in the Polish education system, whereby middle school was due to disappear and be replaced with a new system with 8 years of elementary school and four years of high school. The authors surveyed teachers in middle and primary schools with two messages, either prevention or promotion.
}
\references{
M. Roczniewska and Higgins, E.T. (2019). \emph{Messaging organizational change: How regulatory fit relates to openness to change through fairness perceptions}, Journal of Experimental Social Psychology, 85, \url{https://doi.org/10.1016/j.jesp.2019.103882}

M. Roczniewska and Higgins, E.T. (2019). \emph{Messaging organizational change: How regulatory fit relates to openness to change through fairness perceptions}, Journal of Experimental Social Psychology, 85, \url{https://doi.org/10.1016/j.jesp.2019.103882}
>>>>>>> update2
}
\keyword{datasets}
