% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{RJ74}
\alias{RJ74}
\title{Rosen and Jerdee (1974)}
\format{
A data frame with 4 rows and 3 variables:
\describe{
  \item{\code{gender}}{[factor] gender, either \code{male} or \code{female}}
  \item{\code{decision}}{[factor] outcome, either \code{promote} or \code{hold file}}
  \item{\code{admin}}{[factor] complexity of job, either \code{simple} or \code{complex}}
  \item{\code{count}}{[integer] number of occurences of combination}
}
}
\usage{
RJ74
}
\description{
Data are extracted from Table 1. They consist of assessment of bank supervisors on the promotion of hypothetical employees to the task of branch manager. The purpose of the study was to check  discrimination for promotion.
}
\references{
Rosen, B., and Jerdee, T. H. (1974). \emph{Influence of sex role stereotypes on personnel decisions.} Journal of Applied Psychology, 59(\bold{1}), 9–14. \url{https://doi.org/10.1037/h0035834}
}
\keyword{datasets}
