% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{RT22_S2}
\alias{RT22_S2}
\title{de la Rosa and Tully (2022), Study 2}
\format{
A data frame with 401 rows and 8 variables:
\describe{
  \item{\code{gender}}{[factor] gender of participant}
  \item{\code{age}}{[numeric] age}
  \item{\code{decision}}{[integer] To what extent did the praticipant try to make decisions as they would in real life}
  \item{\code{nexpdec}}{[integer] number of expensive decisions}
  \item{\code{totspend}}{[numeric] total spending}
  \item{\code{nover}}{[integer] number of overdraft charges (35 each, in addition to total spending)}
  \item{\code{condition}}{[factor] experimental condition, either \code{week} for weekly pay or \code{day} for daily transfer}
  \item{\code{subwealth}}{[numeric] subjective wealth index, ranging from low (1) to high (7)}
}
}
\source{
ResearchBox 231, \url{https://researchbox.org/231}, licensed under CC BY 4.0
}
\usage{
RT22_S2
}
\description{
Participants played a life simulator game with different amounts transferred to their "bank account", either on a daily or weekly base.
They had to make 15 binary spending decisions. Excess expenses resulted in overdraft fees of 35 dollars. The authors recorded the total amount spent, the sequential decisions and the number of overdraft charges.
}
\references{
Wendy De La Rosa, Stephanie M Tully (2022). The Impact of Payment Frequency on Consumer Spending and Subjective Wealth Perceptions, \emph{Journal of Consumer Research}, 48(\bold{6}), 991-1009, \doi{10.1093/jcr/ucab052}
}
\keyword{datasets}
