% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{RT22_S4}
\alias{RT22_S4}
\title{de la Rosa and Tully (2022), Study 4}
\format{
A data frame with 528 rows and 6 variables:
\describe{
  \item{\code{gender}}{[factor] gender of participant}
  \item{\code{age}}{[integer] age}
  \item{\code{payfreq}}{[factor] experimental factor for payment frequency, either daily or weekly}
  \item{\code{expensfreq}}{[factor] experimental factor for payment frequency, either daily or weekly}
  \item{\code{subjectivewealth}}{[numeric] average of six Likert scales on subjective wealth, on a scale of 0 (low) to 100 (high)}
  \item{\code{preduncertain}}{[numeric] average of four Likert scales on prediction uncertainty, on a scale of 0 to 100}
}
}
\source{
ResearchBox 231, \url{https://researchbox.org/231}, licensed under CC BY 4.0
}
\usage{
RT22_S4
}
\description{
Study on the frequency of expenses vs payments for the perception of subjective wealth and predicted uncertainty. Only people who did not participate in a similar study or passed attention checks are included.
}
\references{
Wendy De La Rosa, Stephanie M Tully (2022). The Impact of Payment Frequency on Consumer Spending and Subjective Wealth Perceptions, \emph{Journal of Consumer Research}, 48(\bold{6}), 991-1009, \doi{10.1093/jcr/ucab052}
}
\keyword{datasets}
