% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{RT22_S5}
\alias{RT22_S5}
\title{de la Rosa and Tully (2022), Study 5}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1120 rows and 10 columns.
}
\source{
ResearchBox 231, \url{https://researchbox.org/231}, licensed under CC BY 4.0
}
\usage{
RT22_S5
}
\description{
Study 5 is a replication of Study 2, with additionalmeasures of intertemporal discount rate, financial literacy,
Participants played a four-week life simulator and received either 1400 biweekly (starting on the first Friday), or 140 daily on weekdays. This ensured different balances even with the same spending decisions. Participants had to choose 28 daily decisions. There was no overdraft fee in this study.
\describe{
  \item{\code{gender}}{[factor] gender of participant}
  \item{\code{age}}{[integer] age}
  \item{\code{subjectivewealth}}{[numeric] average of six Likert scales on subjective wealth, on a scale of 0 (low) to 100 (high)}
  \item{\code{preduncertain}}{[numeric] average of four Likert scales on prediction uncertainty, on a scale of 0 to 100}
  \item{\code{valence}}{[integer] reported valence from the 3-item self-assessment Manikin scale}
  \item{\code{arousal}}{[integer] reported arousal from the 3-item self-assessment Manikin scale}
  \item{\code{power}}{[integer] reported power from the 3-item self-assessment Manikin scale}
  \item{\code{condition}}{[factor] experimental condition, either \code{bi-weekly} for biweekly pay or \code{dayly} for daily transfer}
  \item{\code{nexpdec}}{[integer] number of expensive decisions}
  \item{\code{totspend}}{[numeric] total spending}
}
}
\references{
Wendy De La Rosa, Stephanie M Tully (2022). The Impact of Payment Frequency on Consumer Spending and Subjective Wealth Perceptions, \emph{Journal of Consumer Research}, 48(\bold{6}), 991-1009, \doi{10.1093/jcr/ucab052}
}
\keyword{datasets}
