% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SE24_S2B}
\alias{SE24_S2B}
\title{DATASET_TITLE}
\format{
A data frame with 401 rows and 8 variables:
\describe{
  \item{\code{age}}{[integer] age of person}
  \item{\code{gender}}{[factor] either \code{male}, \code{female} or \code{other}}
  \item{\code{cues}}{[factor] number of cues, either \code{one} or \code{two}}
  \item{\code{picture}}{[factor] whether a picture is present in the stimulus, either \code{absent} or \code{present}}
  \item{\code{like}}{[numeric] Likert scale for liking of scent 'I like the fragrance of clementines', ranging from strongly disagree (1) to strongly agree (5)}
  \item{\code{involve}}{[numeric] involvement with the product item}
  \item{\code{evaluation}}{[numeric] average of Likert scales for evaluation, with values ranging from strongly disagree (1) to strongly agree (5)}
  \item{\code{olfactory}}{[numeric] average of Likert scales for olfactory imagery item, with values ranging from strongly disagree (1) to strongly agree (5)}
}
}
\source{
ResearchBox 2061, \url{https://researchbox.org/2061}, licensed under CC BY 4.0
}
\usage{
SE24_S2B
}
\description{
DATASET_DESCRIPTION
}
\references{
Sharma, Varun; Estes, Zachary (2024). Seeing is Smelling: Pictures Improve Product Evaluations by Evoking Olfactory Imagery, \emph{International Journal of Research in Marketing}, 41(\bold{2}), 282–307. \doi{10.1016/j.ijresmar.2024.02.001}
}
\keyword{datasets}
