% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SE24_S3}
\alias{SE24_S3}
\title{Sharma and Estes (2024), Study 3}
\format{
A data frame with 798 rows and 9 variables:
\describe{
  \item{\code{id}}{[factor] participant identifier}
  \item{\code{age}}{[integer] age of person}
  \item{\code{gender}}{[factor] either \code{male}, \code{female} or \code{other}}
  \item{\code{like}}{[numeric] Likert scale for liking of scent 'I like the fragrance of lemons', ranging from strongly disagree (1) to strongly agree (5)}
  \item{\code{involve}}{[numeric] involvement with the product item}
  \item{\code{enfas}}{[numeric] Evaluation of the Need for Active Smell scale}
  \item{\code{lemon}}{[factor] within-subject factor, one of \code{whole} or \code{cut}}
  \item{\code{evaluation}}{[numeric] average of Likert scales for evaluation, with values ranging from strongly disagree (1) to strongly agree (5)}
  \item{\code{olfactory}}{[numeric] average of Likert scales for olfactory imagery item, with values ranging from strongly disagree (1) to strongly agree (5)}
}
}
\source{
ResearchBox 2061, \url{https://researchbox.org/2061}, licensed under CC BY 4.0
}
\usage{
SE24_S3
}
\description{
Long format repeated measures on the mediation effect of olfactory imagery on product evaluation. 399 participants were recruited on prolifics and were asked to rate cleaning products with images of whole and cut lemons.
}
\references{
Sharma, Varun; Estes, Zachary (2024). Seeing is Smelling: Pictures Improve Product Evaluations by Evoking Olfactory Imagery, \emph{International Journal of Research in Marketing}, 41(\bold{2}), 282–307. \doi{10.1016/j.ijresmar.2024.02.001}
}
\keyword{datasets}
