% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SSVB21_S2}
\alias{SSVB21_S2}
\title{van Stekelenburg et al. (2021), Study 2}
\format{
A data frame with 442 rows and 3 variables:
\describe{
  \item{\code{prior}}{[integer] negative of prior belief score}
  \item{\code{post}}{[integer] post experiment score for veracity of claim}
  \item{\code{condition}}{[factor] experimental condition, one of \code{Boost}, \code{BoostPlus} and \code{consensus}}
}
}
\source{
Open Science Foundation, \url{https://osf.io/hua8v/}, licensed under CC-BY Attribution 4.0 International
}
\usage{
SSVB21_S2
}
\description{
The study aimed to assess changes in opinion of skeptics when faced with evidence of scientific consensus through manipulation of experimental condition explaining how to assess a news claim. The subject in question was Genetically Engineered Food (GEF).
}
\references{
Stekelenburg, A. van, Schaap, G., Veling, H., & Buijzen, M. (2021). \emph{Boosting understanding and identification of scientific consensus can help to correct false beliefs.} Psychological Science, 32(\bold{10}), 1549–1565. \url{https://doi.org/10.1177/09567976211007788}
}
\keyword{datasets}
