% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{STC21_SS5}
\alias{STC21_SS5}
\title{Sharma, Tully, and Cryder (2021), supplementary study 5}
\format{
A data frame with 1501 rows and 3 variables:
\describe{
  \item{\code{purchase}}{[factor] experimental condition for purchase, either \code{discretionary} or non-discretionary (\code{need})}
  \item{\code{debttype}}{[factor] experimental condition for label of debt type, either \code{loan} or \code{credit}}
  \item{\code{likelihood}}{[double] average score of likelihood in/ interest of using offer to make the purchase (1-9 scale)}
}
}
\source{
Research Box 111, \url{https://researchbox.org/111}, unspecified license.
}
\usage{
STC21_SS5
}
\description{
Supplementary study 5 of Sharma, Tully, and Cryder (2021).
The data consist of all observations which correctly identified an attention check.
The author tested "whether participants are more interested in borrowed funds when those funds are labeled as credit rather than a loan". The preregistration is available at \url{https://aspredicted.org/blind.php?x=uv2f5p}
}
\references{
Sharma, E., Tully, S. and C. Cryder (2021). \emph{Psychological Ownership of (Borrowed) Money}, Journal of Marketing Research, \bold{58}(3), \url{https://doi.org/10.1177/0022243721993816}
}
\keyword{datasets}
