% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{teller}
\alias{teller}
\title{Teller data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{\code{course}}{[factor] whether a formal training was offered (either \code{no} or \code{yes})}
  \item{\code{nweeks}}{[factor] number of weeks of training one-to-one with an experienced clerk}
  \item{\code{error}}{[double] monthly error in balance account}
}
}
\source{
Berger, P.D, Maurer, R.E. and G.B. Celli (2018). Experimental Design: With Application in Management, Engineering, and the Sciences, 2nd edition, Springer.
}
\usage{
teller
}
\description{
Fictional data from Berger et al. (2018) discussed  in Example 6.4 (First United Federal Bank of Boston). One hunder bank tellers were offered one-on-one training with an experienced clerk for a certain number of weeks on the job, possibly in addition to formal training period. The goal of the study is to determine which combination is most efficient at reducing the monthly error in balance (in dollars).
}
\keyword{datasets}
