% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{workhours}
\alias{workhours}
\title{Work hours}
\format{
A data frame with 320 rows and 3 variables:
\describe{
  \item{\code{id}}{[factor] employee identifier}
  \item{\code{week}}{[factor] week of timing}
  \item{\code{time}}{[double] time spent on HEC premises}
}
}
\usage{
workhours
}
\description{
Simulated data for a one-way random effect model on employee attendance.
}
\author{
Leo Belzile
}
\keyword{datasets}
