% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{AA21}
\alias{AA21}
\title{Amirabdolahian and Ali-Adeeb (2021)}
\format{
A data frame with 1169 rows and 3 variables:
\describe{
  \item{\code{latency}}{[double] latency of N170 EEG signal for Fz electrode}
  \item{\code{epoch}}{[factor] ordering of picture}
  \item{\code{stimulus}}{[factor] image shown, one of real image (\code{R}), a one-pass generative adversarial network generate image with slightly happy face (\code{GAN1}) and a two-pass GAN with an extremely happy person (\code{GAN2})}
  \item{\code{id}}{[factor] participant identifier}
}
}
\source{
Saeed Amirabdolahian, personal communication, licensed under CC BY-NC-SA 4.0
}
\usage{
AA21
}
\description{
A study on the attitude towards real and computer generated image of people smiling. The response is the latency for a particular electrode located at the back of the brain.
}
\details{
The data were collected as part of a course project during Winter 2021 (TECH 80747A Information Technologies and Neuroscience), which took place during the Covid19 pandemic. Each student participant had his own apparatus, which explains the large between-participant measurement error.
}
\references{
S. Amirabdolahian and R. Ali-Adeeb (2021). \emph{Measuring face images visual fidelity in neurophysiological level: a study on facial expression manipulating technologies}, technical report, HEC Montreal.
}
\keyword{datasets}
