% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{BL22_E}
\alias{BL22_E}
\title{Brucks and Levav (2022), eyetracker data}
\format{
A data frame with 270 rows and 6 variables:
\describe{
 \item{\code{id}}{[factor] identifier of dyad}
 \item{\code{writer}}{[logical] whether the person was randomly assigned to be the writer}
  \item{\code{cond}}{[factor] experimental condition, either \code{video} for videoconferencing or \code{f2f} for face-to-face}
  \item{\code{partner_time}}{[double] time spent looking at partner (in seconds)}
  \item{\code{task_time}}{[integer] time spent looking at the task computer (in seconds)}
  \item{\code{room_time}}{[double] time spent looking at room (in seconds)}
}
}
\source{
Research Box 282, \url{https://researchbox.org/282}, licensed under CC BY 4.0
}
\usage{
BL22_E
}
\description{
Brucks and Levav (2022), eyetracker data
}
\details{
The authors measured via an eyetracker the time spent gazing at the room, at the other partner and at the task during their lab experiment
}
\examples{
\dontrun{
library(lme4)
mod <- lmer(task_time ~ cond + writer + (1|id),
            data = BL22_E)
summary(mod)
}
}
\references{
Brucks, M.S. and J. Levav (2022). \emph{Virtual communication curbs creative idea generation}. Nature 605, 108–112. \doi{10.1038/s41586-022-04643-y}
}
\keyword{datasets}
