% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{E24_S5}
\alias{E24_S5}
\title{Evangelidis (2024), Study 5}
\format{
A data frame with 1202 rows and 5 variables:
\describe{
  \item{\code{perception}}{[factor] perception of customer, either \code{fair} or \code{unfair}}
  \item{\code{firmdec}}{[factor] firm decision, either \code{price increase} or \code{product downsizing}}
  \item{\code{costs}}{[factor] reason behind change to costs or product size, whether costs \code{increased} or stayed the \code{same}}
  \item{\code{gender}}{[factor] gender of respondant}
  \item{\code{age}}{[integer] age of respondant}
}
}
\source{
Research Box 1032, \url{https://researchbox.org/1032}, licensed under CC BY 4.0
}
\usage{
E24_S5
}
\description{
This study considered a simple between-subject balanced design where shrinkflation is compared to price increases, depending on market conditions. The response is the perception of customers on the proposed change.
}
\references{
I. Evangelidis (2024). Frontiers: Shrinkflation Aversion: When and Why Product Size Decreases Are Seen as More Unfair than Equivalent Price Increases. \emph{Marketing Science}, 43(\bold{2}), 280–288. \doi{10.1287/mksc.2023.0269}
}
\keyword{datasets}
