% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{HOSM22_E3}
\alias{HOSM22_E3}
\title{Hatano et al. (2022), Experiment 3}
\format{
A data frame with 126 rows and 4 variables:
\describe{
  \item{\code{id}}{[factor] subject identifier}
  \item{\code{waiting}}{[factor] experimental condition for waiting time, either \code{short} or \code{long}}
  \item{\code{ratingtype}}{[factor] experimental within-subject factor, either \code{prediction} or \code{experience}}
  \item{\code{imscore}}{[double] response, intrinsic motivation score}
}
}
\usage{
HOSM22_E3
}
\description{
The data consist in a two by two mixed analysis of variance. The authors studied engagement and enjoyment from waiting tasks, and "potential effects of time interval on the underestimation of task motivation by manipulating the time for the waiting task". The waiting time was randomly assigned to either short (3 minutes) or long (20 minutes).
}
\references{
Hatano, A., Ogulmus, C., Shigemasu, H., & Murayama, K. (2022). \emph{Thinking About Thinking: People Underestimate How Enjoyable and Engaging Just Waiting Is}. Journal of Experimental Psychology: General, 151(\bold{12}), 3213–3229 \url{http://dx.doi.org/10.1037/xge0001255}
}
\keyword{datasets}
