% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{HOSM22_E4}
\alias{HOSM22_E4}
\title{Hatano et al. (2022), Experiment 4}
\format{
A data frame with 80 rows and 4 variables:
\describe{
  \item{\code{id}}{[factor] subject identifier}
  \item{\code{ratingtype}}{[factor] experimental within-subject factor, either \code{prediction} or \code{experience}}
  \item{\code{imscore}}{[double] response, intrinsic motivation score}
  \item{\code{irtime}}{[integer] interim rating time minus 10 minutes}
}
}
\usage{
HOSM22_E4
}
\description{
The authors examined the enjoyment and engagement from waiting tasks, doing nothing. The participants were asked to predict their levels of enjoyment and engagement during and after the task. The authors fixed a linear mixed model with an interaction between rating type and interim rating time; the latter is centered by subtracting ten minutes (the mode is 9).
}
\references{
Hatano, A., Ogulmus, C., Shigemasu, H., & Murayama, K. (2022). \emph{Thinking About Thinking: People Underestimate How Enjoyable and Engaging Just Waiting Is}. Journal of Experimental Psychology: General, 151(\bold{12}), 3213–3229 \url{http://dx.doi.org/10.1037/xge0001255}
}
\keyword{datasets}
