% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{JZBJG22_E2}
\alias{JZBJG22_E2}
\title{Jordan et al. (2022), Experiment 2}
\format{
A data frame with 582 rows and 7 variables:
\describe{
  \item{\code{condition}}{[factor] experimental condition, one of \code{video} or \code{no video}}
  \item{\code{order}}{[factor] order of the questions, \code{dying first} or \code{pilot first}}
  \item{\code{conf_dying}}{[integer] Likert scale from 0 (not at all confident to 100 (very confident) for the question: "How confident are you that you would be able to land the plane without dying"}
  \item{\code{conf_pilot}}{[integer] Likert scale from 0 (not at all confident to 100 (very confident) for the question: "How confident are you that you would be able to successfully land the plane as well as a pilot could"}
  \item{\code{expertise}}{[integer] Likert scale ranging from no expertise (\code{1}) to a great deal of expertise (\code{5}) answering the question "How much expertise do you think is involved in landing a plane"}
  \item{\code{ease_imagining}}{[integer] Likert scale ranging from not at all difficult (\code{1}) to very difficult (\code{5}) for the answer to the question "How difficult was it for you to imagine attempting to land the plane."}
  \item{\code{gender}}{[factor] gender of participant, one of \code{man}, \code{woman} or other (\code{gender diverse})}
}
}
\source{
Research Box 511, \url{https://researchbox.org/511}, licensed under CC BY 4.0
}
\usage{
JZBJG22_E2
}
\description{
The authors studied how confidence in people skill to land a plane (with no experience) and it was affected by watching a trivially related video showing a landing.
}
\references{
Jordan, K., R. Zajac, D. Bernstein, C. Joshi and M. Garry (2022). \emph{Trivially informative semantic context inflates people's confidence they can perform a highly complex skill}, Royal Society Open Science,\bold{9}, 211977, \url{http://doi.org/10.1098/rsos.211977}
}
\keyword{datasets}
