% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{PCSCFL24_S3}
\alias{PCSCFL24_S3}
\title{Poirier et al. (2024), Study 3}
\format{
A data frame with 302 rows and 3 variables:
\describe{
  \item{\code{subset}}{[factor] experimental comparison}
  \item{\code{condition}}{[factor] social presence dummy index}
  \item{\code{choice}}{[integer] choice between two products}
}
}
\source{
Sara-Maude Poirier, personal communication, distributed under CC BY-NC-SA 4.0
}
\usage{
PCSCFL24_S3
}
\description{
Data for Study 3, which assigned every participant to a pairwise comparison, indexed by \code{condition}. They had to select one treadmill from two models (counterbalanced), with images showing one of the
}
\note{
The authors seemingly compared the interaction in a two-way contingency table using Pearson chi-square statistic.
}
\examples{
data_novshigh <- subset(PCSCFL24_S3, subset == "no vs high", select = c(choice, condition))
(tab_novshigh <- table(data_novshigh))
chisq.test(tab_novshigh, correct = FALSE)
}
\references{
Poirier, S.-M., Cosby, S., Sénécal, S., Coursaris, C. K., Fredette, M. et Léger, P.-M. (2024). \emph{The impact of social presence cues in social media product photos on consumers’ purchase intentions}. Journal of Business Research, 185, 114932. \doi{10.1016/j.jbusres.2024.114932}.
}
\keyword{datasets}
