% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{words}
\alias{words}
\title{Effects of verbal feedback}
\format{



}
\source{
Keppel and Wickens (2004), Table 21.1, p.466

Keppel and Wickens (2004), Table 21.1, p.466
}
\usage{
words

words
}
\description{
Fictional balanced 3 by 3 by 2 design with five replicatoins from Keppel and Wickens (Chapter 21).
The scenario considered 'effects of verbal feedback given during the acquisition of different types of learning material on memory tested one week later. The response is the number of words remembers from a list of vocabulary words in the lab.
}
\references{
Keppel, G., & Wickens, T. D. (2004). \emph{Design and analysis: A researcher’s handbook.}, 4th edition, Pearson Prentice Hall.

Keppel, G., & Wickens, T. D. (2004). \emph{Design and analysis: A researcher’s handbook.}, 4th edition, Pearson Prentice Hall.
}
\keyword{datasets}
